Partial-order concurrency: permit data races
Data races are undefined behaviour, and we shouldn't be giving a
semantics for these. Instead, omit write-serialisation constraints when
at least one write is non-atomic.

This is work in progress as more debugging is to be done and we need to
figure out the right way to conditionally add assertions (for a
--data-race-check option).