Add a `justfile` with basic support for building CBMC and executing Rust API tests.
This is a rudimentary file, added as a base for easy running of the
Rust API tests locally and online - as it allows for literally one-step
build + test.

Over time I intend for this to be used for two purposes:

1. Streamline building of CBMC for existing and new contributors
2. Streamline build infrastructure in CI.

Also adding mysefl as a codeowner for the file just so that I can be
notified about any changes made to it (assuming I'm the person with
the most `just` knowledge for now).