Fix broken multiline reference
The `<!--` `-->` hack to continue a reference overmultiple lines no
longer works in Doxygen 1.9.1. This commit avoids the need for a hack by
fitting the reference on a single line. The class name `ai_baset` can be
safely ommitted from the reference, because this block of doxygen is
already assoicated with the `ai_baset` class.