Functionality for initializing shadow memory
Allocates shadows for a given original memory
for each declared field. The shadows are registered
in the shadow_memory_state (part of goto_symex_statet
to consider scoping of local shadows).

In the next step this functionality is used to
allocate shadow memory for the original memory
allocations encountered when symexing a
goto model.