Fully remove rewrite_index
The deprecating PR #7030 was merged on 2022-08-11, which is more than 6
months ago. This cleanup also ensures there are no dependencies on the
object file left (an object file that the build wouldn't clean out
anymore for the source was removed from its `Makefile` in the
deprecating commit 2d68a2c52a7).