Return result of `collect_deref_expr` function by value
This allows for `deref_expr_r` to be made constant, which makes it clear
that the set is rebuilt for each assign instruction and is built in one
place only. This also makes it clear that the result of the
`collect_deref_expr` function is its return value.

There is no performance cost to this refactor as Return Value
Optimization (RVO) applies in this case.