Avoid confusing "__CPROVER_{r,w,rw}_ok is not declared" warning
With 4b05d48 there no longer is a (pseudo-) forward declaration
available, making the C front-end warn about this built-in function not
being declared. As this is polymorphic, we shouldn't be type checking it
as a function in the first place, and instead should give it separate
treatment as we do with other polymorphic built-ins.