ITE and binary relations are constant when all operands are constants
This expands the definition of is_constantt to cover if-then-else as
well as binary relations and logic or bit operations. In many, but not,
cases the simplifier will have turned these into simpler constants. The
included regression test is derived from Rust code, where the
intermediate representation (together with constant propagation) will
yield such comparisons.