Add error handling for exit code of Chocolatey
Powershell does not automatically exit a script when an `.exe` returns a
non-successful exit code. This is due to a lack of exit code
standardisation across programs for the Windows platform. Therefore we
need to check for the particular exit code of an `.exe` in order to
report it to the github actions platform. Note that exit codes from
`.exe` invocations are placed in `$LastExitCode` and exit codes from
Powershell cmdlets such as `Write-Output` are placed in the separate
`$?` variable.