Remove dependency on corrosion.
We no longer need to keep the C++/Rust builds in sync, as the Rust build
now depends on the static library libcprover.x.y.z.a being present anywhere
in the user's computer, thus corrosion brings no benefits, and ties us down
to keeping two build systems in sync when we don't need to, along with waisting
time building the rust project twice when we want to run the tests.

To build the Rust project, please do from the src/libcprover-rust/ directory,
using `cargo`, based on the information present on its own directory.