[REFACTOR] Move utility functions into a new module `ffi_util`.
That should help with API self-explainability, along with better
organisation in the future, pulling all utility functions related
to translating types across the FFI bridge into a single module.

Along with this, we also kill the `unsafe extern "Rust"` which is
kinda useless - we aim for this project to be a one way bridge
of C++ -> Rust, and we don't intend to export Rust functions or
types to the C++ end.