Remove link to device-behaviour documentation
I would expect the doxygen website to work with either `\ref
Device-behaviour` to link to the section or `\ref modeling-mmio.md` to
link to the file. However the documentation build system was not able to
resolve either of these links at the time of writing this commit.

I tried using `[modeling-mmio.md](../cprover-manual/modeling-mmio.md)`
which should work when viewing the page as markdown locally or through
githubs UI. However this introduced a failure on the doxygen checking PR
job, for which it was non-trivial to add to the list of known warnings,
as the warning message contained an absolute path.