C front-end: ensure array type updates are consistent
We must not end up in a situation where the symbol's array type is
different from the type applied to symbol expressions in code. With the
previous approach, support-function generation would alter the type
after typechecking of code had already been completed. Instead, do as
the C standard says: if the type is incomplete _at the end of a
translation unit_, the (implicit) initializer makes it a one-element
array.

Fixes: #7608