Do not error on undef extern linker symbols
goto-cc currently parses the linker script given by the -T flag to find
addresses for variables that are extern-declared in C files but not
given a definition. This is because some variables are defined in the
linker script, not in C code, and reading these definitions from the
linker script is necessary to verify the C program.

Prior to this commit, goto-cc would error out if some extern-defined
variables remained undefined even after reading the linker script. There
is actually a valid use case for having undefined symbols even after
linking, see this (search for "Undefined symbol to cause link failure"):

    https://lore.kernel.org/all/5548CC6D.1040708@citrix.com/T/

This commit turns the error into a warning and synthesizes a null
pointer for such symbols.