Add support for reading previous deprecated goto binary version
This commit is intended to be reverted as part of cbmc major release
version 6. The idea is to temporarily include support for the old
version whilst it is needed for transitional reasons. The long term plan
is to remove it in order to remove the maintenance burden of supporting
the old version.