CMake build system: support system-provided MiniSat
Some operating systems provide packaged versions of MiniSat. When
building distribution packages it may then be preferable to use the
pre-installed library instead of building the solver object from
scratch. With this patch it is now possible to select "system-minisat2"
as a solver, alongside any other solvers that may be built from scratch.

The required support for external libraries also addresses #7641.

Fixes: #7641