Adding initial version of `goto-inspect`.
This allows to inspect a goto-binary (this means showing the goto-functions,
for now at least).

The reason this binary is helpful is that different tools (CBMC, goto-instrument, etc)
perform instrumentation before showing the goto-functions, and there's a need
to be able to inspect certain properties of a goto-binary without doing instrumentation.