add a 'language feature' facility for goto programs
This adds a facility to track the language features used by a goto program.
The features are stored in the value of a symbol in the symbol table part of
a goto model.  The default, when not specificd, is 'true', i.e., we
conservatively assume that the feature might be in use unless we know
otherwise.

Checks for three language features not supported by goto symex are added
(assembler, function pointers, vectors).