Add capability to export symex-ready-goto programs
This allows CBMC to dump the goto-program when it is in core-goto
form in a file on disk (when it is just before SYMEX, and all transformations
have been applied).

This allows CBMC to dump the goto-program when it is in symex-ready-goto
form in a file on disk (i.e. just before SYMEX, and all transformations
have been applied).