Variadic args/va_arg: handle unexpected va_list types
Our regression test has `void*` lists when we'd expected `void**`. This,
however, went unnoticed as we only tried to access the first element.
Attempts to access the second element would have been misaligned, and
failed an invariant as of 35cc503 (we don't handle pointer arithmetic
over void pointers in the back-end anymore).

We now make sure that we always increment by size-of-`void*`.

Fixes: #7706