Value sets: make pointers nondet only once
When `get_value_set_rec` discovers a nondet symbol it will consider the
pointer pointing to any of the known objects (as of 37896707577). It
suffices to do this once for each run of `get_value_set`, even when
multiple nondet symbols are encountered while traversing an expression.

This reduces the symex time on the test of #7357 from 930 seconds to 404
seconds.