Change is_constantt logic to enumerate not-constant expressions
Rather than adding more-and-more expressions to the enumeration just
call out those that are _not_ constant ((nondet) symbols and side-effect
expressions as well as any access that may be out of bounds).

With the additional const-ness we need to make Array_UF22 a bit more
complicated (less constant) to continue to demonstrate the problem.