goto-symex can safely propagate with_exprt
We would only deem with_exprt constant when all their operands are
constant, which means that the array operand must be an array or
array_of expression, and both update indices as well as update values
must be constant. In most such cases the simplifier would rewrite the
with_exprt anyway. The exception to this is an index that cannot
immediately be evaluated to a numeric constant, as is the case with,
e.g., pointer_object expressions.

Fixes: #3095