Allow simplification of singleton interval.
Before this change, an expression like `value >= 255 && value <= 255`
would fail to be simplified to `value == 255` by the expression
simplifier.

While the simplification itself is useful, the *lack* of it was causing
some misbehaviour in one of our backends (the old SMT one under
`src/solvers/smt2`) because the lack of simplification resulted
in less constraints being propagated to the solver, missing some assertions
around the expression that was using `value` as the size of an array.