Lower enum_tags to its enum underlying type.
In order to support expressions with type `c_enum_tag_typet` in the
incremental SMT2 solver, a new `lower_enum` has been added.
Such function lowers the type of each expression node in the AST with
type `c_enum_tag_type` to the underlying type of the `c_enum_typet` that
the tag points to.