Add support for dynamic object sizes in the old SMT backend
The old SMT backend simply ignored the sizes of dynamically sized
objects instead of actually defining the size. This allowed the solver
to just choose whichever size it liked in order to reach an outcome
of SAT, rather than applying any bounds which had been applied to the
size. Actually translating the size where the expression is non-constant
should fix various observable behaviors related to object sizes and
bounds.