Add lookup for non-lowered exprt in get_identifier
This fixes a problem with looking up the value of an assertion, which
resulted in the properties not being properly updated after the
property decider had run.
This was because assertion expressions are given a handle symbol that is
queried to the solver. The decision procedure was trying to resolve the
handle from a lowered expression and not the original that was saved in
the map not receiving the correct symbol back.