Expand options pack in `symex-ready-goto/chain.sh`
This would previously been left as a single unexpanded argument
regardless of the number of arguments in the `.desc` file. For example
`goto_bmc main.goto.symex_ready ''` in the case of no arguments
specified. This should be expanded to match the number of arguments in
the `options` variable. For example `goto_bmc main.goto.symex_ready` in
the case of no arguments or `goto_bmc main.goto.symex_ready --argument1
--argument2` in the case of two arguments in the desc file.