Improved checks and fixed duplicate_per_byte
Function `duplicate_per_byte` now has a PRECONDITION checking that the
`init_byte_expr` is either a boolean or a bitvector of maximum size 8.

Also the computation of the number of duplication correctly accounts the
case when the destination is not divisible by 8. In this case the input
is duplicated one extra (including the sign bit if any) time and then
truncated to the output size.