Fix combination of `new-smt-backend` and `gcc-only` test tags
Tests which were tagged with both `new-smt-backend` and `gcc-only` would
previously be executed on Windows with Visual Studio test jobs. The
`gcc-only` tag is used when tests cover features which are not supported
by Visual Studio. Therefore running them with Visual Studio is a an
error on the part of the test framework, rather than representative of
an issue with cbmc. Propagating `gcc_only_string` to the addition of the
`new-smt-backend` test profile will switch off running these tests for
the Visual Studio platform only.