cbmc-incr-smt2 tests: remove misguided assertion
There is no reason an integer couldn't have the same value as a pointer
(when converted to an integer type). This assertion might accidentally
hold on platforms where sizeof(int) != sizeof(int *) given our current
pointer model, but this is an implementation detail that we must not
rely on. The assertion certainly doesn't hold when sizeof(int) ==
sizeof(int *).