Fix byte update of struct/array elements
A key part of this is factoring out code that was almost-the-same from
lower_byte_update_array_vector and lower_byte_update_struct, but
actually had subtle bugs as documented in regression test included in
this commit.

For consistency, do all offset calculations in bits rather than flipping
back and forth between bits and bytes.