Fix duplicate_per_byte when output is float
When the output of `duplicate_per_byte` is `float`, the init_value gets
converted to `float` and not reinterpreted.
This in combination with shift operation causes the return value to be
different than the replication of init_value.

This commit solves this issue by applying a reinterpret cast to float
using byte_extract.
Also this commit re-enables the shadow memory regression test that was
failing due to duplicate_per_byte to a float value.