Fix duplicated type names in generated parsers
Roughly speaking, the one definition rule of C++ forbids multiple
different definitions of the same thing within the final program. But
violations of the rule are not required to be errored or warned about by
build tool chains. The multiple generated parsers within the cbmc
code base each have their own definitions of the `yyalloc` and
`yysymbol_kind_t` types, which is a violation of the rule. This
violation is not picked up by when building using the current tool
chains and build options. However it is picked up if we use Link Time
Optimisation (LTO). Fixing the rule violation is required in preparation
for switing on LTO, but would be a good idea even without it.