Add simplify_expr step to duplicate_per_byte
The function `duplicate_per_byte` contains a more efficient implementation
when `init_expr` is a constant.
However this feature is never used as the shadow memory module almost
always wraps `init_expr` with a typecast to the type of the shadow
memory.

This commit applies a simplification step in `duplicate_per_byte`,
calling `simplify_expr`, so that the casts may be removed and the more
efficient implementation will be used.