Make constant_propagation_08 test platform agnostic
We had a varying number of simplifications as platforms with sizeof(int)
!= sizeof(ssize_t) would yield 2 simplifications, and those with
matching bit widths only a single one. The extra simplification referred
to the cast in the index operand (on platforms with sizeof(int) !=
sizeof(ssize_t)), which we can avoid by making the index a long-type
constant in the first place.