Refactor three-way-merge so that it makes more sense
Given that the operation is taking the diff between function start and
function end and applying it to a base, you would think that it should only
have three domains.  It previously had four and then copied the entire
calling domain over the "this" domain, which was not otherwise used.
This patch simplifies this and removes that oddity.