Fix encoding of structs which contain a single struct field
In the case where a struct contains a single struct typed field, then
the construction of the outer class is encoded into the construction of
the inner class which then itself needs to be encoded. Therefore the
result of the encoding itself may need encoding. This general solution
will also avoid further variations on the same issue such as
constructing unions containing structs, or selecting struct fields from
a union etc.