Fix book regression tests handling of smt test runs
The `CMakeLists.txt` for the `book-examples` directory appears to be
based on a copy of the one from the `regression/cbmc` directory.
Unfortunately the first parameter to `add_test_pl_profile` needs to be
unique and it has not been updated to a unique name for the new file.
This was resulting in the existing tests not being run.