Avoid adding useless goto blocks in if with return
When the parsed AST is transformed in GOTO language the if conditions
are translated in: `if (!c) goto else; then_branch; goto endif;
else: else_branch; endif: nop` and `if (!c) goto endif; then_branch;
endif: nop`.

This commit avoids to add the `goto endif` command and the `endif` label
if the `then_branch` ends with a `return` or a `goto` statement.

The benefit is that the `--cover location` command returns a
more precise result as unreachable code blocks that are added by CBMC's
frontend are not marked as unreachable.