Conditionally simplify use of if-then-else conditions
When symbolically executing (and dereferencing) statements like
`*p = *p + 1;` where `p` may point to more than a single object we ended
up with nested if-then-else expressions on the right-hand side that
would use the very same conditions. In this particular case, we should
propagate the conditions into nested if-then-else expressions. This
simplification may be expensive, so we don't enable it in general.