Replace `new-smt-backend` tag with inverted `no-new-smt` tag
This makes it somewhat more straight forward to find tests where the new
smt backend is not currently run. It also helps with testing with the
new smt support on new tests by default.

The tags in the `.desc` files are flipped using a one-off python script.
I am including it here below in the commit message body rather than in a
separate commit as it is intended for a singular usage and it may be
more straight forward to review the script than the `.desc` file
updates.

```
import os
import sys
from pathlib import Path

def usage():
    print("Usage " + sys.argv[0] + " [directory]")
    print("Processes all test tags within a [directory].")

def update_file(file_path) -> bool:
    with file_path.open("rt") as file:
        lines = file.readlines()
    if len(lines) == 0:
        print("File " + file_path + " is empty")
        return False
    if "CORE" not in lines[0]:
        return False
if " new-smt-backend" not in lines[0] and " no-new-smt" not in
lines[0]:
        lines[0] = lines[0].replace("\n", " no-new-smt\n")
    else:
        lines[0] = lines[0].replace(" new-smt-backend", "")
    with file_path.open("wt") as file:
        file.writelines(lines)
    return True

def main() -> int:
    if len(sys.argv) != 2:
        usage()
        return os.EX_USAGE
    directory = Path(sys.argv[1])
    if not directory.is_absolute():
        directory = Path(os.getcwd()) / directory
    print("Running against directory " + os.fspath(directory))
    if not directory.is_dir():
        print("Error - This is not a directory.")
        return os.EX_IOERR
    files = directory.glob("*/*.desc")
    tests_updated = 0
    for file_path in files:
        if not file_path.is_file():
print(os.fspath(file_path) + " was expected to be a file but
is "
                                         "not a file")
            continue
        if update_file(file_path):
            tests_updated += 1
    if tests_updated == 0:
        print("Error - No tests updated")
        return os.EX_NOTFOUND
    print(str(tests_updated) + " tests updated.")
    return os.EX_OK

if __name__ == '__main__':
    sys.exit(main())
```