Restore post-order simplification of byte_extract(c ? a : b, ...)
With 5d294910ff it was assumed that this step was fully taken care of by
pre-order simplification passes. Other post-order simplification rules,
however, may newly introduce such expressions, making it necessary that
we _also_ do this in post-order traversal.

Absence of this rule leads to doubling of memory consumption when trying
to prove aws_cryptosdk_keyring_trace_copy_all in aws-encrption-sdk-c.