Fix libstdc++-not-found error
See
https://github.com/diffblue/cbmc/actions/runs/6856483568/job/18643716705?pr=6749
for one example of the failure: cargo-induced clang runs (via the cc
crate) weren't finding the C++ library (which ought to be libc++ and not
libstdc++ on this MacOS target). This appears to be caused by the
minimum build target being too low (where the cc crate is the one
setting that minimum). Override that by setting an environment variable.
See
https://github.com/rust-lang/cc-rs/blob/2d6a3b2119cf5eacc01e1f2877e064a7aede7819/src/lib.rs#L3497C52-L3497C76
for the Rust code implementing the logic.