Use `if constexpr` in `smt_function_application_termt::indices`
With C++17's `if constexpr` feature we can now remove two alternate
function templates and just put the alternate implementations in line.
This puts the specialisation into the outer function, so that it is
specialised with the correct implementation for the given `functiont`,
rather than dispatching based on the `std::true_type` /
`std::false_type` overloads.