Add hopping through declarations on jump into scope
When a goto which jumps into a scope is converted, the declarations
for that scope need to be passed through. It needs to be the same
declaration in both execution paths, so that symex will correctly merge
the two different states when generating the phi functions.

The requirement for executing the same declaration whether entering the
scope through a regular route vs a goto into the middle of the block is
acheived by setting up a chain of hops where -
 * Original goto hops to first declaration.
 * After first declaration we hop to the next declaration if entering
   via a goto.
 * Afer the final declaration we hop to the originally specified label.

The identifiers in the existing `symex_should_filter_value_sets` test
need to be incremented, because the instance being checked is now the
second time the variable is added to the scope.