CSmith test script: avoid a need for argv modelling
With additional checks turned on as of #8093, we failed CSmith tests
with (legitimate) pointer property failures in `strcmp`. These are
caused by trying to compare `argv[1]` to a string. As we do not model
`argv` in these tests, `argv[1]` was not a valid pointer. This fix just
removes the string comparison, which is only used for turning on/off
debug output in test execution.