Remove @NlightNFotis from `CODEOWNERS`.
With my tenure at Diffblue coming to an end, and a vision of myself
working in different domains going forward, I don't think there's
going to be much time to spare working on CBMC on the side. To that
end, I believe it's fair to the project to remove myself from the
`CODEOWNERS` file so that nothing is pending my approval in the future.

It has been an honour and a pleasure working on the project, and
I'm happy with what we've achieved these years.