Goto crossing scopes: fix scope tree entry of conditions
We must convert the condition of an if/then/else first to make sure any
declarations produced by converting the condition have a suitable
scope-tree node. Previously, gotos in either branch of the if-then-else
were led to believe that declarations resulting from condition were not
yet in scope. The feature from #8091 would, thus, result in a spurious
loop back to the declaration to put it in scope before following the
goto edge.