Support comparing arrays of non-constant size with array_equal
Using simple type equality is too strict for arrays of non-constant
size: we introduce a fresh symbol for the size of each dynamically
allocated object of non-constant size. Consequently, two dynamically
allocated arrays of non-constant size will never pass type equality
checking, even when their underlying sizes are the same.

We now explicitly compare the sizes of the two arrays passed to
array_equal when types are not trivially equal.

Fixes: #8176