simplify extractbits_exprt representation
The current representation of extractbits_exprt stores both the upper and
lower indices of the range that is to be extracted, and their difference
plus one in form of the width of the type of the expression.

This removes the upper index, as it can be deduced from the lower index and
the width of the result.  The key benefit is reducing burden on the user of
the class, who a) doesn't have to remember which index comes first, and b)
doesn't have to do the calculation of the upper index.