test.pl now reports both directory and descriptor file name
The test.pl script identifies tests by the directory and the name of the
descriptor file name.  Historically, the directory was sufficient, since the
descriptor file name used to be hardwired to test.desc.  This changes the
output to include both the path and the name of the file, which removes the
need to look for directory and file name in two different places.