introduce 'fatal assertions'
This introduces a variant of ASSERT instructions that are fatal when they
are refuted.  Execution paths through fatal assertions that are refuted are
undefined.  Assertions that (otherwise) pass and that are reachable from a
refuted fatal assertion are now reported as UNKNOWN.

The motivating use-case for fatal assertions is undefined behavior in
languages such as C/C++ or Rust.

Noi attempt is made to modify the outcome of ASSERT instructions that are
refuted owing to a trace through a refuted fatal assertion.  The computation
of the trace is not robust enough.