embedd files using hex literals instead of string literals
String constants are limited to 64k, which requires splitting the file with
the compiler headers.

This switches the embedding to use an array of hex literals, which have no
size limit.

There is a prospect to use the #embed preprocessor command in C++26, moves
the conversion from file to hex literal array into the compiler's
preprocessor.