GOTO conversion: create temporaries with minimal scope
GOTO conversion introduces temporaries when cleaning expression, e.g.,
removing side effects. We previously considered them to have block scope
as they only were marked dead when the block containing them was left.
This, however, can be a much larger range of instructions than for what
instructions they actually need to be live for. As a consequence, GOTO
conversion frequently deemed it necessary to introduce declaration hops
for we had goto instructions that would jump over the declaration of the
temporary, but still within the block that contained that temporary (and
well after the last actual use of that temporary).

This PR now largely (with the exception of compound literals, which
yield temporaries that must have block scope) removes the side effect
that creating temporaries had on scope tracking. Instead, methods
explicitly return the list of temporaries in need of cleanup.

This avoids performance penalties seen when trying to upgrade Kani to
CBMC version 6. Kani makes extensive use of statement expressions, which
are one case of instructions that yield a temporary that needs to be
cleaned up as soon as possible.