SMT2 back-end: report error when invoking solver fails
Our previous attempt at reporting errors was dead code for `run` never
returns a negative value. When the executable isn't found the return
code was 1, which is also a legitimate return code when `z3` runs
successfully. Thus, use emptiness of `stderr` output as the success
criterion, and propagate the messages produced by `run` on stderr to the
error-logging output.

Fixes: #8362