Do not define project(CBMC ...) twice to fix CMake failures
`project(P, ...)` sets the CMake variablee `P_SOURCE_DIR` to whichever
directory that CMake file is in.  We happen to rely on the value of
`CBMC_SOURCE_DIR` in several places.  Invoking `project(P, ...)` twice
in different directories for the same value of `P` will cause
hard-to-reproduce behaviour.  Even though this duplication was in place
ever since, 7949cac5f7a19, we apparently got lucky until a few days ago.
Now, however, we see sporadic failures on GitHub runners, such as
https://github.com/diffblue/cbmc/actions/runs/10677152028/job/29591809621.
GitHub's runner image notes do not point out any recent change to CMake,
but it might as well be changes to the kernel's scheduler.