bitvector_typet: set width from mp_integer
This adds C-style getter/setter methods to bitvector_typet for the width of
the vector.

The setter takes an mp_integer, which checks that the width fits within
std::size_t.  This removes the burden of performing this check from the
caller, and allows callers to avoid potential arithmetic overflows by using
mp_integer.