Use same CMake minimum version in DownloadProject configuration
We already have a top-level minimum requirement for CMake >= 3.8, so
don't need to specify a lower requirement in a file included from the top
level.

Having a requirement as low as 2.8.2 broke building on newer version of
CMake, which error with "Compatibility with CMake < 3.5 has been removed
from CMake." (seen in GitHub runners with Ubuntu 24.04).