prophecy_r_or_w_ok_exprt lowering: adjust for dynamic/static objects
We do not need to create comparisons to dead and/or deallocated when
handling a dynamic object (which will never be marked "dead") or a
static object (which will neither be marked "dead" nor "deallocated").
This avoids unnecessary branches when the prophecy_r_or_w_ok_exprt can
now be simplified to a constant.