Feature to --retrace a single control flow path
Adds a feature to retrace and verify a single control flow path. The path
can be given using `--retrace` as a string of 0s and 1s. From a very brief
evaluation test, it seems to be faster than bounded loop unwinding. Only
that it might be challenging to find a path to retrace. The 0s and 1s in
the input retrace string follow the goto decisions in cbmc's internal goto
program representation. During symbolic execution, files and line numbers
for each decision are logged to the console, so the interactive use should
be intuitive.

Co-authored-by: Yosephine Setiawan <agatayosephine@gmail.com>