run(..., std::ostream &, ...) with pipe
run(...) has a variant that that accepts an std::ostream for the stdout
redirection.  This replaces the use of popen(...) in this variant by a
standard combination of pipe/fork/exec.

The benefit is this change is that the use of a shell is avoided, which is
difficult to secure owing to the parsing of the command by the shell.