Fix support for `with_exprt` with more than 3 operands
A `with_exprt` needs to have at least 3 operands, but can encode
multiple updates when using more operands (where any further operands
need to come in pairs of two: an index/member and a new value). We
already support this in several places, but were still missing support
in others. This led to wrong verification results in Kani as recent
changes in CBMC make increasing use of the value set (which is among
those fixed in this commit).